:recycle: [#433] -- extract utilities from date input group
These can be used for the datepicker locale-aware bits
